+++ sub.cpp
@@ -1,4 +1,4 @@
-int sub(int a, int b)
+double sub(double a, double b)
 {
     return a - b;
 }
